python/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
openai
scikit-learn
numpy
pandas
python-dotenv
pydantic
aiohttp
umap-learn
click
rich
anthropic
google-generativeai
transformers
torch
aiosqlite
asyncpg
tiktoken
langdetect
psutil
unicodedata2